Refactor: 생성일: 날짜 포맷팅 맞추기(연,월,일)
#115
